Got almost all code working on qrisk male as starting point. Need to find where Zoe's performance comparison plots are and add them in.
