+++ PREDICT/Triggers.py
@@ -281,7 +281,7 @@ def __SPCCalculateControlLimits(model, input_data, startCLDate, endCLDate, warni
 
             errors_by_date = createCLdf.groupby(model.dateCol).apply(CalculateError)
             model.mean_error = errors_by_date.mean()
-            std_dev_error = errors_by_date.std()
+            std_dev_error = errors_by_date.std() / np.sqrt(len(errors_by_date))
             
             model.u2sdl = model.mean_error + warningSDs * std_dev_error
             model.u3sdl = model.mean_error + recalSDs * std_dev_error
